// src/classes/Table/types.d.ts
export declare type DynamoDBTypes = 'string' | 'boolean' | 'number' | 'bigint' | 'list' | 'map' | 'binary' | 'set';
export declare type DynamoDBKeyTypes = 'string' | 'number' | 'bigint' | 'binary';